:sparkles: Création de projects sans vérifs

+++ src/api/main/java/isen/projet_dp_api/repository/ProjectRepository.java
@@ -3,5 +3,5 @@ package isen.projet_dp_api.repository;
 import isen.projet_dp_api.model.dao.ProjectDAO;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<ProjectDAO, Long> {
+public interface ProjectRepository extends JpaRepository<ProjectDAO, Integer> {
 }
